Renamed onUpdate and onUpdateAndDoNow methods; Created a separate class for KeyboardShortcuts component
