+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/components/drawer/SettingsView.kt
@@ -19,17 +19,10 @@
 package org.wycliffeassociates.otter.jvm.workbookapp.ui.components.drawer
 
 import javafx.application.Platform
-import javafx.geometry.HPos
-import javafx.geometry.Pos
-import javafx.scene.Node
-import javafx.scene.control.Label
 import javafx.scene.control.ToggleGroup
-import javafx.scene.layout.HBox
 import javafx.scene.layout.Priority
-import javafx.scene.text.TextAlignment
 import org.kordamp.ikonli.javafx.FontIcon
 import org.kordamp.ikonli.materialdesign.MaterialDesign
-import org.wycliffeassociates.otter.jvm.controls.Shortcut
 import org.wycliffeassociates.otter.jvm.controls.button.SelectButton
 import org.wycliffeassociates.otter.jvm.controls.dialog.confirmdialog
 import org.wycliffeassociates.otter.jvm.controls.styles.tryImportStylesheet
@@ -286,104 +279,7 @@ class SettingsView : View() {
                     addClass("app-drawer__subtitle")
                 }
 
-                vbox {
-                    addClass("app-drawer__section")
-
-                    gridpane {
-                        vgap = 5.0
-                        constraintsForColumn(0).apply {
-                            hgrow = Priority.ALWAYS
-                        }
-                        constraintsForColumn(1).apply {
-                            halignment = HPos.RIGHT
-                        }
-
-                        add(
-                            Label(messages["action"]).apply {
-                                addClass("app-drawer__subtitle--small")
-                            }, 0, 0
-                        )
-                        add(
-                            Label(messages["shortcut"]).apply {
-                                addClass("app-drawer__subtitle--small")
-                            }, 1, 0
-                        )
-
-                        add(makeShortcutAction(messages["focus"]), 0, 1)
-                        add(
-                            makeShortcutBox(makeShortcutButton("tab")), 1, 1
-                        )
-
-                        add(makeShortcutAction(messages["select"]), 0, 2)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton("space"),
-                                makeShortcutButton("enter")
-                            ), 1, 2
-                        )
-
-                        add(makeShortcutAction(messages["navigation"]), 0, 3)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton(icon = FontIcon(MaterialDesign.MDI_ARROW_UP)),
-                                makeShortcutButton(icon = FontIcon(MaterialDesign.MDI_ARROW_DOWN)),
-                                makeShortcutButton(icon = FontIcon(MaterialDesign.MDI_ARROW_LEFT)),
-                                makeShortcutButton(icon = FontIcon(MaterialDesign.MDI_ARROW_RIGHT))
-                            ), 1, 3
-                        )
-
-                        add(makeShortcutAction(messages["scrollDown"]), 0, 4)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton("PgDn"),
-                                makeShortcutButton(icon = FontIcon(MaterialDesign.MDI_ARROW_DOWN))
-                            ), 1, 4
-                        )
-
-                        add(makeShortcutAction(messages["scrollUp"]), 0, 5)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton("PgUp"),
-                                makeShortcutButton(icon = FontIcon(MaterialDesign.MDI_ARROW_UP))
-                            ), 1, 5
-                        )
-
-                        add(makeShortcutAction(messages["goBack"]), 0, 6)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton(Shortcut.GO_BACK.value.displayText)
-                            ), 1, 6
-                        )
-
-                        add(makeShortcutAction(messages["addVerseMarker"]), 0, 7)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton(Shortcut.ADD_MARKER.value.displayText)
-                            ), 1, 7
-                        )
-
-                        add(makeShortcutAction(messages["recordStop"]), 0, 8)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton(Shortcut.RECORD.value.displayText)
-                            ), 1, 8
-                        )
-
-                        add(makeShortcutAction(messages["playPauseSource"]), 0, 9)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton(Shortcut.PLAY_SOURCE.value.displayText)
-                            ), 1, 9
-                        )
-
-                        add(makeShortcutAction(messages["playPauseTarget"]), 0, 10)
-                        add(
-                            makeShortcutBox(
-                                makeShortcutButton(Shortcut.PLAY_TARGET.value.displayText)
-                            ), 1, 10
-                        )
-                    }
-                }
+                add(KeyboardShortcuts())
             }
         }
     }
@@ -425,29 +321,6 @@ class SettingsView : View() {
 
         viewModel.showChangeLanguageSuccessDialogProperty.onChange {
             Platform.runLater { if (it) successDialog.open() else successDialog.close() }
-        }
-    }
-
-    private fun makeShortcutAction(action: String): Label {
-        return Label(action).apply {
-            addClass("app-drawer__text")
-        }
-    }
-
-    private fun makeShortcutBox(vararg shortcuts: Node): Node {
-        return HBox().apply {
-            spacing = 5.0
-            alignment = Pos.TOP_RIGHT
-            children.addAll(shortcuts)
-        }
-    }
-
-    private fun makeShortcutButton(str: String? = null, icon: Node? = null): Node {
-        return Label().apply {
-            addClass("app-drawer__shortcut")
-            text = str?.replace("+", " + ")
-            graphic = icon
-            textAlignment = TextAlignment.CENTER
         }
     }
 }
