+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/screens/RecordResourcePage.kt
@@ -27,7 +27,7 @@ import org.wycliffeassociates.otter.jvm.controls.Shortcut
 import org.wycliffeassociates.otter.jvm.controls.breadcrumbs.BreadCrumb
 import org.wycliffeassociates.otter.jvm.controls.styles.tryImportStylesheet
 import org.wycliffeassociates.otter.jvm.utils.getNotNull
-import org.wycliffeassociates.otter.jvm.utils.onUpdateAndDoNow
+import org.wycliffeassociates.otter.jvm.utils.onChangeAndDoNowWithListener
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.NavigationMediator
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.RecordResourceViewModel
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.viewmodel.WorkbookDataStore
@@ -88,7 +88,7 @@ class RecordResourcePage : View() {
     override fun onDock() {
         tabs.forEach { recordableTab ->
             recordableTab.bindProperties()
-            recordableTab.recordableProperty.onUpdateAndDoNow {
+            recordableTab.recordableProperty.onChangeAndDoNowWithListener {
                 addRemoveTabs(recordableTab, it)
             }.let { recordableTab.recordableListener = it }
         }
